changed system_status variable to 32bit

+++ Ricardo_OS/Ricardo_OS/src/Logging/messages.cpp
@@ -2,20 +2,20 @@
 
 
 //global system status variable intialized at 0
-uint16_t system_status = 0;
+uint32_t system_status = 0;
 
-void new_message(uint16_t msg,String reason){
+void new_message(uint32_t msg,String reason){
 
     system_status = system_status + msg;
 
 
     //todo detect when in usb debug mode and print reason over serial. Maybe worth logging reason to sd card??
-    
+
 
 
 };
 
-void delete_message(uint16_t msg){
+void delete_message(uint32_t msg){
 
     system_status = system_status - msg;
